quizplay/requirements.txt
setuptools >= 21.0.0
gunicorn==22.0.0
flask-marshmallow==1.1.0
flask-cors==4.0.1
connexion[swagger-ui]==3.1.0
connexion[flask]
connexion[uvicorn]
redis==5.2.0
hashring==1.5.1